# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a REST API for saving films you want to watch.",
        "title": "Watchlist API",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/collections/{collection_id}/films": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get a list of films from collection by collection ID. It also returns metadata.\nYou must have permissions to get this collection.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collectionFilms"
                ],
                "summary": "Get films from collection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Collection ID",
                        "name": "collection_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Specify the desired `page`",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Specify the desired `page size`",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting by `film_id`, `added_at`. Use `-` for desc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.CollectionFilmsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/collections/{collection_id}/films/{film_id}": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get the film from collection by ID. You must have permissions to get this collection.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collectionFilms"
                ],
                "summary": "Get film from collection by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Collection ID",
                        "name": "collection_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Film ID",
                        "name": "film_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.CollectionFilmResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Update the film in the collection by ID`s. You must have the permissions to update collection.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collectionFilms"
                ],
                "summary": "Update film in collection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Collection ID",
                        "name": "collection_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Film ID",
                        "name": "film_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New information about the film in the collection",
                        "name": "film",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.CollectionFilmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.CollectionFilmResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Add a film to the collection. You must have rights to get the film and update the collection.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collectionFilms"
                ],
                "summary": "Add film to collection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Collection ID",
                        "name": "collection_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Film ID",
                        "name": "film_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/swagger.CollectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/collections/{collection_id}/films/{films_id}": {
            "delete": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Delete the film from the collection by ID. You must have the permissions to update collection.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collectionFilms"
                ],
                "summary": "Delete film from collection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Collection ID",
                        "name": "collection_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Film ID",
                        "name": "film_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "filters.Metadata": {
            "type": "object",
            "properties": {
                "current_page": {
                    "description": "The current page number.",
                    "type": "integer",
                    "example": 2
                },
                "first_page": {
                    "description": "The first page number, usually 1.",
                    "type": "integer",
                    "example": 1
                },
                "last_page": {
                    "description": "The last page number based on total records and page size.",
                    "type": "integer",
                    "example": 3
                },
                "page_size": {
                    "description": "The number of items per page.",
                    "type": "integer",
                    "example": 5
                },
                "total_records": {
                    "description": "The total number of records available.",
                    "type": "integer",
                    "example": 15
                }
            }
        },
...
        "models.Collection": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "description": "Timestamp when the collection was created.",
                    "type": "string",
                    "example": "2024-09-04T13:37:24.87653+05:00"
                },
                "description": {
                    "description": "Description of the collection; optional, up to 500 characters.",
                    "type": "string",
                    "maxLength": 500,
                    "example": "This is description"
                },
                "id": {
                    "description": "Unique identifier for the collection.",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "Name of the collection; required, between 3 and 100 characters.",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "My collection"
                },
                "updated_at": {
                    "description": "Timestamp when the collection was last updated.",
                    "type": "string",
                    "example": "2024-09-04T13:37:24.87653+05:00"
                },
                "user_id": {
                    "description": "Identifier of the user who created the collection.",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.CollectionFilm": {
            "type": "object",
            "properties": {
                "added_at": {
                    "description": "Timestamp when the film was added to the collection.",
                    "type": "string",
                    "example": "2024-09-04T13:37:24.87653+05:00"
                },
                "collection_id": {
                    "description": "Identifier of the collection.",
                    "type": "integer",
                    "example": 1
                },
                "film_id": {
                    "description": "Identifier of the film.",
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "description": "Timestamp when the association was last updated.",
                    "type": "string",
                    "example": "2024-09-04T13:37:24.87653+05:00"
                }
            }
        },
...
        "swagger.CollectionFilmRequest": {
            "type": "object",
            "properties": {
                "added_at": {
                    "type": "string",
                    "example": "2024-09-04T13:37:24.87653+05:00"
                }
            }
        },
        "swagger.CollectionFilmResponse": {
            "type": "object",
            "properties": {
                "collection_film": {
                    "$ref": "#/definitions/models.CollectionFilm"
                }
            }
        },
        "swagger.CollectionFilmsResponse": {
            "type": "object",
            "properties": {
                "collection_films": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CollectionFilm"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/filters.Metadata"
                }
            }
        },
...
        "swagger.CollectionResponse": {
            "type": "object",
            "properties": {
                "collection": {
                    "$ref": "#/definitions/models.Collection"
                }
            }
        },
...
        "swagger.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "some kind of error"
                }
            }
        },
...
        "swagger.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "some kind of success message"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWTAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}